All energy values become uint64_t to avoid overflow. Add a rough implementation to getEnergyOptimisticNWH(), good enough for now. Add the action type TRIED_AND_FAILED so that failed actions can be tracked and included in energy costs. Print heap stats correctly. Add log points EVENT_ENERGY_REQUIRED_TOTAL_NWH, EVENT_HEAP_LEFT, EVENT_STACK_LEFT, EVENT_HEAP_MIN_LEFT and APPLICATION_LOG_VERSION becomes 4. Put heap, heap minimum and stack values LOGX(). Bring in new version of log client which allows logs to also be directly printed under the define LOG_PRINT_IMMEDIATELY.

+++ source/actions/act_cellular.h
@@ -52,7 +52,7 @@
  *                          + RRC Wait time @ 98mA (assumed 6 seconds)
  *                          + RRC Release 185 uWh
  */
-#define CELLULAR_R410_ENERGY_TX_NWH(x) (((unsigned int) x) * 25 + 17500 + 588000 + 185000)
+#define CELLULAR_R410_ENERGY_TX_NWH(x) (((uint64_t) x) * 25 + 17500 + 588000 + 185000)
 
 /** The energy required, in nWh, for the N2xx modem to transmit x bytes.
  *
@@ -62,7 +62,7 @@
  * Send X Bytes = 0.05894 * X + 11.54 uWh
  *                + RRC wait time @ 48mA (assumed 6 seconds)
  */
-#define CELLULAR_N2XX_ENERGY_TX_NWH(x) (34000 + ((unsigned int) x) * 59 + 11540 + 288000)
+#define CELLULAR_N2XX_ENERGY_TX_NWH(x) (34000 + ((uint64_t) x) * 59 + 11540 + 288000)
 
 /**************************************************************************
  * TYPES
